+++ src/render_stats.js
@@ -1,0 +1,7 @@
+var renderStats = function(stats, options) {
+  console.log('  All TODOs:      ' + stats.all);
+  console.log('  Reviewed TODOs: ' + stats.reviewed);
+  // TODO: use stats.files for statistics per file
+}
+
+module.exports = renderStats;
